+++ tests/turns/handlers/playerTurnHandler.fixes.test.js
@@ -14,11 +14,7 @@ import HumanTurnHandler from '../../../src/turns/handlers/humanTurnHandler.js';
 describe('HumanTurnHandler Constructor', () => {
   let mockLogger;
   let mockTurnStateFactory;
-  let mockCommandProcessor;
   let mockTurnEndPort;
-  let mockPlayerPromptService;
-  let mockCommandOutcomeInterpreter;
-  let mockSafeEventDispatcher;
   let mockGameWorldAccess;
   let mockTurnContextBuilder;
   let mockTurnStrategyFactory; // <-- Changed
@@ -36,11 +32,7 @@ describe('HumanTurnHandler Constructor', () => {
     mockTurnStateFactory = {
       createInitialState: jest.fn().mockReturnValue(mockInitialState),
     };
-    mockCommandProcessor = {};
     mockTurnEndPort = {};
-    mockPlayerPromptService = {};
-    mockCommandOutcomeInterpreter = {};
-    mockSafeEventDispatcher = {};
     mockGameWorldAccess = {};
     mockTurnContextBuilder = {
       build: jest.fn(({ actor }) => ({
@@ -72,13 +64,8 @@ describe('HumanTurnHandler Constructor', () => {
   const getValidDependencies = () => ({
     logger: mockLogger,
     turnStateFactory: mockTurnStateFactory,
-    commandProcessor: mockCommandProcessor,
     turnEndPort: mockTurnEndPort,
-    promptCoordinator: mockPlayerPromptService,
-    commandOutcomeInterpreter: mockCommandOutcomeInterpreter,
-    safeEventDispatcher: mockSafeEventDispatcher,
-    turnContextBuilder: mockTurnContextBuilder,
-    turnStrategyFactory: mockTurnStrategyFactory, // <-- Changed
+    turnStrategyFactory: mockTurnStrategyFactory,
     turnContextBuilder: mockTurnContextBuilder,
     gameWorldAccess: mockGameWorldAccess,
   });
@@ -120,43 +107,11 @@ describe('HumanTurnHandler Constructor', () => {
     );
   });
 
-  it('should throw an error if commandProcessor is not provided', () => {
-    const deps = getValidDependencies();
-    delete deps.commandProcessor;
-    expect(() => new HumanTurnHandler(deps)).toThrow(
-      'HumanTurnHandler: commandProcessor is required'
-    );
-  });
-
   it('should throw an error if turnEndPort is not provided', () => {
     const deps = getValidDependencies();
     delete deps.turnEndPort;
     expect(() => new HumanTurnHandler(deps)).toThrow(
       'GenericTurnHandler: turnEndPort is required'
-    );
-  });
-
-  it('should throw an error if promptCoordinator is not provided', () => {
-    const deps = getValidDependencies();
-    delete deps.promptCoordinator;
-    expect(() => new HumanTurnHandler(deps)).toThrow(
-      'HumanTurnHandler: promptCoordinator is required'
-    );
-  });
-
-  it('should throw an error if commandOutcomeInterpreter is not provided', () => {
-    const deps = getValidDependencies();
-    delete deps.commandOutcomeInterpreter;
-    expect(() => new HumanTurnHandler(deps)).toThrow(
-      'HumanTurnHandler: commandOutcomeInterpreter is required'
-    );
-  });
-
-  it('should throw an error if safeEventDispatcher is not provided', () => {
-    const deps = getValidDependencies();
-    delete deps.safeEventDispatcher;
-    expect(() => new HumanTurnHandler(deps)).toThrow(
-      'HumanTurnHandler: safeEventDispatcher is required'
     );
   });
